leetcode/src/test/java/org/issxjl2015/algorithm/Lc_79_单词搜索/SolutionTest.java
package org.issxjl2015.algorithm.Lc_79_单词搜索;

import org.junit.Assert;
import org.junit.Test;

public class SolutionTest {

    @Test
    public void test1() {
        Solution solution = new Solution();
        char[][] board = {{'A', 'B', 'C', 'E'}, {'S', 'F', 'C', 'S'}, {'A', 'D', 'E', 'E'}};
        String word = "ABCCED";
        boolean ans = solution.exist(board, word);
        Assert.assertTrue(ans);
    }

    @Test
    public void test2() {
        Solution solution = new Solution();
        char[][] board = {{'A', 'B', 'C', 'E'}, {'S', 'F', 'C', 'S'}, {'A', 'D', 'E', 'E'}};
        String word = "SEE";
        boolean ans = solution.exist(board, word);
        Assert.assertTrue(ans);
    }

    @Test
    public void test3() {
        Solution solution = new Solution();
        char[][] board = {{'A', 'B', 'C', 'E'}, {'S', 'F', 'C', 'S'}, {'A', 'D', 'E', 'E'}};
        String word = "ABCB";
        boolean ans = solution.exist(board, word);
        Assert.assertFalse(ans);
    }
}
